refactor: openstack, remove SD struct from config, move into openstack discovery package
