+++ config/config.go
@@ -29,6 +29,7 @@ import (
 	"github.com/prometheus/prometheus/discovery/file"
 	"github.com/prometheus/prometheus/discovery/gce"
 	"github.com/prometheus/prometheus/discovery/marathon"
+	"github.com/prometheus/prometheus/discovery/openstack"
 	"github.com/prometheus/prometheus/discovery/triton"
 	"github.com/prometheus/prometheus/discovery/zookeeper"
 
@@ -113,12 +114,6 @@ var (
 
 	// DefaultKubernetesSDConfig is the default Kubernetes SD configuration
 	DefaultKubernetesSDConfig = KubernetesSDConfig{}
-
-	// DefaultOpenstackSDConfig is the default OpenStack SD configuration.
-	DefaultOpenstackSDConfig = OpenstackSDConfig{
-		Port:            80,
-		RefreshInterval: model.Duration(60 * time.Second),
-	}
 
 	// DefaultRemoteWriteConfig is the default remote write configuration.
 	DefaultRemoteWriteConfig = RemoteWriteConfig{
@@ -362,7 +357,7 @@ type ServiceDiscoveryConfig struct {
 	// List of EC2 service discovery configurations.
 	EC2SDConfigs []*ec2.SDConfig `yaml:"ec2_sd_configs,omitempty"`
 	// List of OpenStack service discovery configurations.
-	OpenstackSDConfigs []*OpenstackSDConfig `yaml:"openstack_sd_configs,omitempty"`
+	OpenstackSDConfigs []*openstack.SDConfig `yaml:"openstack_sd_configs,omitempty"`
 	// List of Azure service discovery configurations.
 	AzureSDConfigs []*azure.SDConfig `yaml:"azure_sd_configs,omitempty"`
 	// List of Triton service discovery configurations.
@@ -635,65 +630,6 @@ func (c *KubernetesNamespaceDiscovery) UnmarshalYAML(unmarshal func(interface{})
 		return err
 	}
 	return yamlUtil.CheckOverflow(c.XXX, "namespaces")
-}
-
-// OpenstackSDConfig is the configuration for OpenStack based service discovery.
-type OpenstackSDConfig struct {
-	IdentityEndpoint string            `yaml:"identity_endpoint"`
-	Username         string            `yaml:"username"`
-	UserID           string            `yaml:"userid"`
-	Password         configUtil.Secret `yaml:"password"`
-	ProjectName      string            `yaml:"project_name"`
-	ProjectID        string            `yaml:"project_id"`
-	DomainName       string            `yaml:"domain_name"`
-	DomainID         string            `yaml:"domain_id"`
-	Role             OpenStackRole     `yaml:"role"`
-	Region           string            `yaml:"region"`
-	RefreshInterval  model.Duration    `yaml:"refresh_interval,omitempty"`
-	Port             int               `yaml:"port"`
-
-	// Catches all undefined fields and must be empty after parsing.
-	XXX map[string]interface{} `yaml:",inline"`
-}
-
-// OpenStackRole is role of the target in OpenStack.
-type OpenStackRole string
-
-// The valid options for OpenStackRole.
-const (
-	// OpenStack document reference
-	// https://docs.openstack.org/nova/pike/admin/arch.html#hypervisors
-	OpenStackRoleHypervisor OpenStackRole = "hypervisor"
-	// OpenStack document reference
-	// https://docs.openstack.org/horizon/pike/user/launch-instances.html
-	OpenStackRoleInstance OpenStackRole = "instance"
-)
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface.
-func (c *OpenStackRole) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	if err := unmarshal((*string)(c)); err != nil {
-		return err
-	}
-	switch *c {
-	case OpenStackRoleHypervisor, OpenStackRoleInstance:
-		return nil
-	default:
-		return fmt.Errorf("Unknown OpenStack SD role %q", *c)
-	}
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface.
-func (c *OpenstackSDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	*c = DefaultOpenstackSDConfig
-	type plain OpenstackSDConfig
-	err := unmarshal((*plain)(c))
-	if err != nil {
-		return err
-	}
-	if c.Role == "" {
-		return fmt.Errorf("role missing (one of: instance, hypervisor)")
-	}
-	return yamlUtil.CheckOverflow(c.XXX, "openstack_sd_config")
 }
 
 // RelabelAction is the action to be performed on relabeling.
