+++ discovery/openstack/openstack.go
@@ -16,14 +16,17 @@ package openstack
 import (
 	"context"
 	"errors"
+	"fmt"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack"
 	"github.com/prometheus/client_golang/prometheus"
-
-	"github.com/prometheus/prometheus/config"
+	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/pkg/targetgroup"
+	configUtil "github.com/prometheus/prometheus/util/config"
+	yamlUtil "github.com/prometheus/prometheus/util/yaml"
 )
 
 var (
@@ -37,7 +40,71 @@ var (
 			Name: "prometheus_sd_openstack_refresh_duration_seconds",
 			Help: "The duration of an OpenStack-SD refresh in seconds.",
 		})
+	// DefaultSDConfig is the default OpenStack SD configuration.
+	DefaultSDConfig = SDConfig{
+		Port:            80,
+		RefreshInterval: model.Duration(60 * time.Second),
+	}
 )
+
+// SDConfig is the configuration for OpenStack based service discovery.
+type SDConfig struct {
+	IdentityEndpoint string            `yaml:"identity_endpoint"`
+	Username         string            `yaml:"username"`
+	UserID           string            `yaml:"userid"`
+	Password         configUtil.Secret `yaml:"password"`
+	ProjectName      string            `yaml:"project_name"`
+	ProjectID        string            `yaml:"project_id"`
+	DomainName       string            `yaml:"domain_name"`
+	DomainID         string            `yaml:"domain_id"`
+	Role             OpenStackRole     `yaml:"role"`
+	Region           string            `yaml:"region"`
+	RefreshInterval  model.Duration    `yaml:"refresh_interval,omitempty"`
+	Port             int               `yaml:"port"`
+
+	// Catches all undefined fields and must be empty after parsing.
+	XXX map[string]interface{} `yaml:",inline"`
+}
+
+// OpenStackRole is role of the target in OpenStack.
+type OpenStackRole string
+
+// The valid options for OpenStackRole.
+const (
+	// OpenStack document reference
+	// https://docs.openstack.org/nova/pike/admin/arch.html#hypervisors
+	OpenStackRoleHypervisor OpenStackRole = "hypervisor"
+	// OpenStack document reference
+	// https://docs.openstack.org/horizon/pike/user/launch-instances.html
+	OpenStackRoleInstance OpenStackRole = "instance"
+)
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface.
+func (c *OpenStackRole) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	if err := unmarshal((*string)(c)); err != nil {
+		return err
+	}
+	switch *c {
+	case OpenStackRoleHypervisor, OpenStackRoleInstance:
+		return nil
+	default:
+		return fmt.Errorf("Unknown OpenStack SD role %q", *c)
+	}
+}
+
+// UnmarshalYAML implements the yaml.Unmarshaler interface.
+func (c *SDConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	*c = DefaultSDConfig
+	type plain SDConfig
+	err := unmarshal((*plain)(c))
+	if err != nil {
+		return err
+	}
+	if c.Role == "" {
+		return fmt.Errorf("role missing (one of: instance, hypervisor)")
+	}
+	return yamlUtil.CheckOverflow(c.XXX, "openstack_sd_config")
+}
 
 func init() {
 	prometheus.MustRegister(refreshFailuresCount)
@@ -47,12 +114,12 @@ func init() {
 // Discovery periodically performs OpenStack-SD requests. It implements
 // the TargetProvider interface.
 type Discovery interface {
-	Run(ctx context.Context, ch chan<- []*config.TargetGroup)
-	refresh() (tg *config.TargetGroup, err error)
+	Run(ctx context.Context, ch chan<- []*targetgroup.Group)
+	refresh() (tg *targetgroup.Group, err error)
 }
 
 // NewDiscovery returns a new OpenStackDiscovery which periodically refreshes its targets.
-func NewDiscovery(conf *config.OpenstackSDConfig, l log.Logger) (Discovery, error) {
+func NewDiscovery(conf *SDConfig, l log.Logger) (Discovery, error) {
 	var opts gophercloud.AuthOptions
 	if conf.IdentityEndpoint == "" {
 		var err error
@@ -73,11 +140,11 @@ func NewDiscovery(conf *config.OpenstackSDConfig, l log.Logger) (Discovery, erro
 		}
 	}
 	switch conf.Role {
-	case config.OpenStackRoleHypervisor:
+	case OpenStackRoleHypervisor:
 		hypervisor := NewHypervisorDiscovery(&opts,
 			time.Duration(conf.RefreshInterval), conf.Port, conf.Region, l)
 		return hypervisor, nil
-	case config.OpenStackRoleInstance:
+	case OpenStackRoleInstance:
 		instance := NewInstanceDiscovery(&opts,
 			time.Duration(conf.RefreshInterval), conf.Port, conf.Region, l)
 		return instance, nil
